+++ table/src/main/java/com/mengyunzhi/article/entity/Article.java
@@ -1,4 +1,4 @@
-package com.mengyunzhi.article.repository;
+package com.mengyunzhi.article.entity;
 
 import javax.persistence.*;
 import java.sql.Date;
