+++ table/src/main/java/com/mengyunzhi/article/entity/Contractor.java
@@ -1,4 +1,4 @@
-package com.mengyunzhi.article.repository;
+package com.mengyunzhi.article.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
