+++ table/src/main/java/com/mengyunzhi/article/repository/AttractionRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Attraction;
 import org.springframework.data.repository.CrudRepository;
 
 /**
