+++ table/src/main/java/com/mengyunzhi/article/repository/ContractorRespository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Contractor;
 import org.springframework.data.repository.CrudRepository;
 
 /**
