+++ table/src/main/java/com/mengyunzhi/article/repository/DetailRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Detail;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DetailRepository extends CrudRepository<Detail, Long> {
