+++ table/src/main/java/com/mengyunzhi/article/repository/HotelRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Hotel;
 import org.springframework.data.repository.CrudRepository;
 
 public interface HotelRepository extends CrudRepository<Hotel, Long> {
