+++ table/src/main/java/com/mengyunzhi/article/repository/MaterialRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Material;
 import org.springframework.data.repository.CrudRepository;
 
 public interface MaterialRepository extends CrudRepository<Material, Long> {
