+++ table/src/main/java/com/mengyunzhi/article/repository/ParagraphRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Paragraph;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ParagraphRepository extends CrudRepository<Paragraph, Long> {
