+++ table/src/main/java/com/mengyunzhi/article/repository/PlanRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Plan;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PlanRepository extends CrudRepository<Plan,Long> {
