+++ table/src/main/java/com/mengyunzhi/article/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Long>{
