+++ table/src/test/java/com/mengyunzhi/article/repository/ArticleRepositoryTest.java
@@ -1,5 +1,8 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Article;
+import com.mengyunzhi.article.entity.Contractor;
+import com.mengyunzhi.article.entity.Plan;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
