+++ table/src/test/java/com/mengyunzhi/article/repository/AttractionRepositoryTest.java
@@ -1,12 +1,11 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Attraction;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
