+++ table/src/test/java/com/mengyunzhi/article/repository/ContractorRespositoryTest.java
@@ -1,5 +1,6 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Contractor;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
