+++ table/src/test/java/com/mengyunzhi/article/repository/ParagraphRepositoryTest.java
@@ -1,5 +1,7 @@
 package com.mengyunzhi.article.repository;
 
+import com.mengyunzhi.article.entity.Article;
+import com.mengyunzhi.article.entity.Paragraph;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
